Rename DiagnosticAanlyzerContextHelpers
